+++ src/controls.jl
@@ -1,0 +1,2 @@
+target_marginal(g::Union{SpectralFGN,LAMP}) = copy(Float64.(g.marginal))
+target_marginal(g::FSS) = Float64.(g.rates) ./ sum(g.rates)