+++ test/runtests.jl
@@ -1,5 +1,6 @@
 using Test
 using Random
+using StableRNGs
 using S5
 
 @testset "S5.jl" begin
@@ -7,5 +8,6 @@ using S5
     include("test_pb1.jl")
     include("test_mb1.jl")
     include("test_mb3.jl")
+    include("test_marginal_control.jl")
     include("test_io.jl")
 end